The recent changes from 4.0 to 12.0 as the default ToolsVersion for all project files (as part of the VS2013 work in [1533]) stopped the project files from loading in Visual Studio 2008. The ToolsVersion for all project files is now set to a fake tools version (12.34), and this version needs to match the corresponding tool requirements for each VS version.
To facilitate this, I've added a VSVer.bat to update the registry keys and set the TargetVisualStudioVersion for each VS release. The added version-specific batch files simply call VSVer (VS2008.bat runs VSVer 2008). These batch files need to be run from the corresponding 'Visual Studio 20xx Command Prompt' (opened as an Administrator) before launching a solutions in Visual Studio.

Note that the toolsversion setting was usually provided for a command-line build, effectively overriding the value in the .csproj file. Build batch files have been updated so that this is always the case. However, VS cannot be persuaded to pay attention to any similar tools version setting, and the placement of this attribute on the root element of .csproj file means that it cannot be macroized, so some creativity was required to allow a single project file to open in all versions of VS.